Authentication (#5)
* create user model and validations

* implement user authentication

* implement user authorization for auction and bid actions

+++ api/app/controllers/bids_controller.rb
@@ -1,14 +1,24 @@
 class BidsController < ApplicationController
+    before_action :authenticate_user!
 
     def create
         bid = Bid.new bid_params
         bid.auction = Auction.find params[:auction_id]
-        bid.save
+        bid.user = current_user
+        if bid.save
+            render json: bid.auction
+        else
+            head :conflict
+        end
     end
 
     def destroy
         bid = Bid.find params[:id]
-        bid.destroy
+        if bid.user == current_user 
+            bid.destroy
+        else
+            head :unauthorized
+        end
     end
 
     private
@@ -16,4 +26,5 @@ class BidsController < ApplicationController
     def bid_params
         params.require(:bid).permit :price
     end
+
 end
